feat: Reuse Sponge common code
