+++ backend/common/src/main/java/io/github/_4drian3d/signedvelocity/common/SignedResult.java
