+++ backend/paper/src/main/java/io/github/_4drian3d/signedvelocity/paper/SignedVelocity.java
