+++ backend/sponge/common/build.gradle.kts
@@ -1,0 +1,4 @@
+dependencies {
+    compileOnly("org.spongepowered:spongeapi:8.1.0")
+    implementation(projects.signedvelocityBackendCommon)
+}
