+++ backend/sponge/common/src/main/java/io/github/_4drian3d/signedvelocity/sponge/common/listener/SignedListener.java
@@ -1,0 +1,8 @@
+package io.github._4drian3d.signedvelocity.sponge.common.listener;
+
+import org.spongepowered.api.event.Event;
+import org.spongepowered.api.event.EventListener;
+
+public interface SignedListener<E extends Event> extends EventListener<E> {
+    void register();
+}
