+++ gradle.properties
@@ -1,3 +1,3 @@
 group = io.github._4drian3d
-version = 1.0.0-BETA-3
+version = 1.0.0-RC
 description = Allows you to cancel or modify messages or commands from Velocity without synchronization problems
