--- paper/build.gradle.kts
@@ -1,39 +1,0 @@
-plugins {
-    alias(libs.plugins.runpaper)
-    alias(libs.plugins.pluginyml.paper)
-    alias(libs.plugins.shadow)
-}
-
-dependencies {
-    compileOnly(libs.paper)
-    implementation(projects.signedvelocityCommonBackend)
-}
-
-tasks {
-    clean {
-        // Deletes the directory that is generated by the runPaper plugin
-        delete("run")
-    }
-    build {
-        dependsOn(shadowJar)
-    }
-    shadowJar {
-        archiveBaseName.set("${rootProject.name}-Paper")
-        archiveClassifier.set("")
-        doLast {
-            copy {
-                from(archiveFile)
-                into("${rootProject.projectDir}/build")
-            }
-        }
-    }
-}
-
-paper {
-    name = "SignedVelocity"
-    main = "io.github._4drian3d.signedvelocity.paper.SignedVelocity"
-    authors = listOf("4drian3d")
-
-    foliaSupported = true
-    apiVersion = "1.19"
-}
